chore: update notebook and index
Updated all the packages which caused pandas to raise an exception with how I was trying to remove the outliers from the log_data dataframe. Had to completely rewrite that. Also had to update some other pandas method calls as well. From there I reran the notebook and captured a new html index page for the project page.